csharpプロジェクトの追加

+++ KIGfx/IndexBuffer.cpp
@@ -30,7 +30,7 @@ void IndexBuffer::Set(GLuint primitiveType, const vector<int>& index)
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_Id);
 	glBufferData(GL_ELEMENT_ARRAY_BUFFER, index.size() * sizeof(int), index.data(), GL_STATIC_DRAW);
 	m_PrimitiveType = primitiveType;
-	m_indexSize = index.size();
+	m_indexSize = (GLuint)index.size();
 	Logger::GLError();
 }
